+++ datagen/src/main/java/org/shsts/tinactory/datagen/content/model/MachineModel.java
@@ -22,9 +22,9 @@ import org.shsts.tinactory.content.network.StaticMachineBlock;
 import org.shsts.tinactory.content.network.SubnetBlock;
 import org.shsts.tinactory.core.builder.SimpleBuilder;
 import org.shsts.tinactory.core.electric.Voltage;
-import org.shsts.tinactory.core.network.MachineBlock;
-import org.shsts.tinactory.core.network.PrimitiveBlock;
-import org.shsts.tinactory.core.network.SidedMachineBlock;
+import org.shsts.tinactory.integration.network.MachineBlock;
+import org.shsts.tinactory.integration.network.PrimitiveBlock;
+import org.shsts.tinactory.integration.network.SidedMachineBlock;
 import org.shsts.tinycorelib.datagen.api.builder.IBlockDataBuilder;
 import org.shsts.tinycorelib.datagen.api.context.IDataContext;
 import org.shsts.tinycorelib.datagen.api.context.IEntryDataContext;
