+++ datagen/src/main/java/org/shsts/tinactory/datagen/extra/LitematicaMeta.java
@@ -18,8 +18,8 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.Property;
 import org.shsts.tinactory.content.machine.UnsupportedTypeException;
 import org.shsts.tinactory.core.common.MetaConsumer;
-import org.shsts.tinactory.core.network.PrimitiveBlock;
 import org.shsts.tinactory.core.util.LocHelper;
+import org.shsts.tinactory.integration.network.PrimitiveBlock;
 import org.slf4j.Logger;
 
 import java.io.FileOutputStream;
