+++ mod/src/main/java/org/shsts/tinactory/AllNetworks.java
@@ -6,9 +6,9 @@ import org.shsts.tinactory.api.network.IScheduling;
 import org.shsts.tinactory.content.electric.ElectricComponent;
 import org.shsts.tinactory.content.logistics.LogisticComponent;
 import org.shsts.tinactory.content.logistics.SignalComponent;
-import org.shsts.tinactory.core.builder.SchedulingBuilder;
-import org.shsts.tinactory.core.network.ComponentType;
-import org.shsts.tinactory.core.network.NetworkComponent;
+import org.shsts.tinactory.integration.builder.SchedulingBuilder;
+import org.shsts.tinactory.integration.network.ComponentType;
+import org.shsts.tinactory.integration.network.NetworkComponent;
 import org.shsts.tinycorelib.api.registrate.IRegistrate;
 import org.shsts.tinycorelib.api.registrate.entry.IEntry;
