+++ mod/src/main/java/org/shsts/tinactory/AllRegistries.java
@@ -17,8 +17,8 @@ import org.shsts.tinactory.api.network.IComponentType;
 import org.shsts.tinactory.api.network.IScheduling;
 import org.shsts.tinactory.core.common.SimpleFluid;
 import org.shsts.tinactory.core.metrics.MetricsManager;
-import org.shsts.tinactory.core.network.ComponentType;
 import org.shsts.tinactory.core.network.SchedulingSorter;
+import org.shsts.tinactory.integration.network.ComponentType;
 import org.shsts.tinactory.integration.network.WorldNetworkManagers;
 import org.shsts.tinycorelib.api.blockentity.IEvent;
 import org.shsts.tinycorelib.api.registrate.entry.IEntry;
