+++ mod/src/main/java/org/shsts/tinactory/content/multiblock/LargeBoiler.java
@@ -22,7 +22,7 @@ import org.shsts.tinactory.content.machine.FireBoiler;
 import org.shsts.tinactory.core.machine.Machine;
 import org.shsts.tinactory.core.multiblock.Multiblock;
 import org.shsts.tinactory.core.multiblock.MultiblockInterface;
-import org.shsts.tinactory.core.network.MachineBlock;
+import org.shsts.tinactory.integration.network.MachineBlock;
 import org.shsts.tinycorelib.api.blockentity.IEventManager;
 import org.shsts.tinycorelib.api.registrate.entry.IMenuType;
 import org.slf4j.Logger;
