+++ mod/src/main/java/org/shsts/tinactory/content/multiblock/LargeTurbine.java
@@ -6,8 +6,8 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import org.shsts.tinactory.core.multiblock.Multiblock;
-import org.shsts.tinactory.core.network.MachineBlock;
-import org.shsts.tinactory.core.network.PrimitiveBlock;
+import org.shsts.tinactory.integration.network.MachineBlock;
+import org.shsts.tinactory.integration.network.PrimitiveBlock;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
