+++ mod/src/main/java/org/shsts/tinactory/content/multiblock/MultiblockMeta.java
@@ -22,7 +22,7 @@ import org.shsts.tinactory.core.gui.Layout;
 import org.shsts.tinactory.core.machine.IRecipeProcessor;
 import org.shsts.tinactory.core.multiblock.Multiblock;
 import org.shsts.tinactory.core.multiblock.MultiblockSpec;
-import org.shsts.tinactory.core.network.PrimitiveBlock;
+import org.shsts.tinactory.integration.network.PrimitiveBlock;
 import org.shsts.tinycorelib.api.core.Transformer;
 import org.shsts.tinycorelib.api.registrate.builder.IBlockEntityTypeBuilder;
 import org.shsts.tinycorelib.api.registrate.entry.IEntry;
