+++ mod/src/main/java/org/shsts/tinactory/content/network/FixedMachineBlock.java
@@ -12,7 +12,7 @@ import org.shsts.tinycorelib.api.registrate.entry.IMenuType;
 
 import java.util.function.Supplier;
 
-import static org.shsts.tinactory.core.network.MachineBlock.WORKING;
+import static org.shsts.tinactory.integration.network.MachineBlock.WORKING;
 
 /**
  * Entity block that does not have facing or IO facing, only a working state.
