+++ mod/src/main/java/org/shsts/tinactory/core/multiblock/Multiblock.java
@@ -43,7 +43,7 @@ import java.util.function.Supplier;
 import static org.shsts.tinactory.AllCapabilities.MACHINE;
 import static org.shsts.tinactory.AllEvents.CLIENT_TICK;
 import static org.shsts.tinactory.AllEvents.SET_MACHINE_CONFIG;
-import static org.shsts.tinactory.core.network.MachineBlock.WORKING;
+import static org.shsts.tinactory.integration.network.MachineBlock.WORKING;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
