+++ mod/src/main/java/org/shsts/tinactory/core/multiblock/MultiblockInterfaceBlock.java
@@ -10,7 +10,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.BooleanProperty;
 import org.shsts.tinactory.core.electric.Voltage;
-import org.shsts.tinactory.core.network.SidedMachineBlock;
+import org.shsts.tinactory.integration.network.SidedMachineBlock;
 import org.shsts.tinycorelib.api.registrate.entry.IBlockEntityType;
 import org.shsts.tinycorelib.api.registrate.entry.IMenuType;
